Extract file acceptance partial
